requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
configparser==4.0.2
Django==3.0.2
django-crispy-forms==1.8.1
django-enumfield==2.0.1
django-enumfields==2.0.0
djangorestframework==3.11.0
idna==2.8
importlib-metadata==1.4.0
kombu==4.6.7
more-itertools==8.1.0
mysqlclient==1.4.6
Pillow==7.0.0
pkg-resources==0.0.0
PyMySQL==0.9.3
pytz==2019.3
redis==3.3.11
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
vine==1.3.0
zipp==2.0.1
